refactor: move JSON patch services to generic core/json_patch package
ApiPatchDomainService contained no API-specific logic — only RFC 6902 /
7396 structural validation over two library port interfaces. Rename it to
JsonPatchDomainService and relocate it, along with the JsonPatchService and
JsonMergePatchService port interfaces and their infra implementations, from
core/api/domain_service/ to core/json_patch/domain_service/.

The infra implementations move from infra/domain_service/api/ to
infra/domain_service/json_patch/, and ApiJsonPatchServiceImpl is renamed to
JsonPatchServiceImpl to match the port it implements.
