+++ gravitee-apim-rest-api/gravitee-apim-rest-api-service/src/main/java/io/gravitee/apim/core/json_patch/domain_service/JsonMergePatchService.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.gravitee.apim.core.api.domain_service;
+package io.gravitee.apim.core.json_patch.domain_service;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import io.gravitee.apim.core.exception.ValidationDomainException;
