+++ gravitee-apim-rest-api/gravitee-apim-rest-api-service/src/main/java/io/gravitee/apim/infra/domain_service/json_patch/JsonMergePatchServiceImpl.java
@@ -13,29 +13,28 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.gravitee.apim.infra.domain_service.api;
+package io.gravitee.apim.infra.domain_service.json_patch;
 
 import com.fasterxml.jackson.databind.JsonNode;
-import com.github.fge.jsonpatch.JsonPatch;
 import com.github.fge.jsonpatch.JsonPatchException;
-import io.gravitee.apim.core.api.domain_service.JsonPatchService;
+import com.github.fge.jsonpatch.mergepatch.JsonMergePatch;
 import io.gravitee.apim.core.exception.ValidationDomainException;
-import java.io.IOException;
+import io.gravitee.apim.core.json_patch.domain_service.JsonMergePatchService;
 import org.springframework.stereotype.Service;
 
 /**
  * @author GraviteeSource Team
  */
 @Service
-public class ApiJsonPatchServiceImpl implements JsonPatchService {
+public class JsonMergePatchServiceImpl implements JsonMergePatchService {
 
     @Override
-    public JsonNode applyJsonPatch(JsonNode patch, JsonNode target) throws ValidationDomainException {
+    public JsonNode applyMergePatch(JsonNode patch, JsonNode target) throws ValidationDomainException {
         try {
-            return JsonPatch.fromJson(patch).apply(target);
+            return JsonMergePatch.fromJson(patch).apply(target);
         } catch (JsonPatchException e) {
             throw new ValidationDomainException("Failed to apply patch: " + e.getMessage(), e);
-        } catch (IOException e) {
+        } catch (Exception e) {
             throw new ValidationDomainException("Invalid patch: " + e.getMessage(), e);
         }
     }
